Remove generated manual from repo
I don't have the correct tool versions to generate it locally, so the
one generated directly in the repository is going to be out-of-date
immediately.

We'll ignore the entire pdf directory, too, since all of this data is
generated somehow. We won't remove the other PDF's from the repo until
we have a proper build system in place for them.