Upgrade tools for 9.6.1 release
soar_shuffler.py now uses several environment variables to determine the
locations of various resources. The script has also generally been
cleaned up and modernized.

Look for macOS ARM64 builds to include in the release, and run
macOS_setup.sh with the rest of the Mac setup so that users don't get
security warnings when they try to run the included tools.

Update the release file lists to include C# for all platforms, and fix
a bug causing the Python lib to be missing as well. The 9.6.0 release
was missing several libs.

Delete Soar_Projects_Touchdates.txt, which is not used.