Quote paths in launch scripts
The scripts were not working when placed in a directory with spaces in
the name, so quote path usage everywhere.

Note on batch vs. bash quoting:
* "*" is not expanded in bash when in quotes, while it is in batch
* `set foo="some_path"` leads to the quotes being included in the
variable in batch, while in bash it is ignored. Therefore in batch we
don't use quotes when declaring path variables but we do when using
them, while in bash we use quotes everywhere.